Move igniton schedule control functions into scheduler_ignition_controller.*
