+++ speeduino/scheduler_ignition_controller.h
@@ -31,3 +31,12 @@ void calculateIgnitionAngles(const config2 &page2, const config4 &page4, statuse
  * @param dwellTime Target dwell time
  */
 void setIgnitionChannels(const statuses &current, uint16_t crankAngle, uint16_t dwellTime);
+
+/** @brief Initialize all schedulers to the OFF state */
+void initialiseIgnitionSchedulers(void);
+
+/** @brief Start the timers that drive schedulers  */
+void startIgnitionSchedulers(void);
+
+/** @brief Stop the timers that drive schedulers  */
+void stopIgnitionSchedulers(void);
